+++ cpp/hello/src/callback.rs
@@ -1,13 +1,13 @@
-use libc::{c_float, size_t};
+use libc::{c_int, size_t};
 
 #[no_mangle]
 pub extern "C" fn hello__callback__map_safe(
-    items_in: *const c_float,
-    items_out: *mut c_float,
+    items_in: *const c_int,
+    items_out: *mut c_int,
     item_count: size_t,
-    mapper: extern "C" fn(c_float) -> c_float,
+    mapper: extern "C" fn(c_int) -> c_int,
 ) {
-    let items: &[c_float] = unsafe { std::slice::from_raw_parts(items_in, item_count) };
+    let items: &[c_int] = unsafe { std::slice::from_raw_parts(items_in, item_count) };
     let mapped_items = hello::callback::map(items, |item| mapper(*item));
 
     unsafe { std::ptr::copy_nonoverlapping(mapped_items.as_ptr(), items_out, mapped_items.len()) }
